backend/requirements.txt
Django==5.0.1
graphene-django==3.2.0
django-environ==0.11.2
django-cors-headers==4.3.1
django-graphql-jwt==0.4.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.1